Fixed compiler warning.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@11636 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/tracks/track_object.cpp
@@ -405,7 +405,7 @@ void TrackObject::reset()
         a_node->setRotation(m_init_hpr);
         a_node->setScale(m_init_scale);
         a_node->setLoopMode(m_is_looped);
-        a_node->setCurrentFrame(a_node->getStartFrame());
+        a_node->setCurrentFrame((float)(a_node->getStartFrame()));
         
         // trick to reset the animation AND also the timer inside it
         a_node->OnAnimate(0);
